requirements.txt
transformers
torch
langdetect
google-generativeai
python-dotenv
fastapi
uvicorn
PyQt5
